Add NOS Idle loop detection and trhottling.
This commit adds NOS2 idle loop detection inside the main loop.
After everything has been stepped but before we go through the next
iteration of the main loop. we test to see if we are out of monitor mode
and if idling is enabled.

If both are true we then test for P = 2 FL = 5 and increment our counter
of idle loop cycles. if we determine that the correct # of idle loop
cycles has elapsed we first check to see that there are no PPs in busy
state. if there are we continue the main loop without sleeping.
This allows us to prioritize I/O and preserve responsiveness.

Finally if no PPs are busy we sleep for the determined amount of time.

The end result is a lot of saved CPU cycles, no racing fans
and 100% cpu available for active I/O and CPU bound jobs.