use nanoseconds on *nix and msec on Windows
Turns out that nanosleep with sub milisecond times works best on Linux
while Sleep() is pretty much the only game in town on Windows.

This patch reverts to using nanosleep with idletime=nsec units
and on Windows we use msec. The idle command has been changed to
reflect this in what it's reporting to the user.